fix: remove stupid int32 <> int conversions in api (#241)

+++ api/internal/service/PostService.go
@@ -79,7 +79,7 @@ func (s *PostService) NewPost(ctx context.Context, currentUser models.PublicUser
 			return errors.New("unable to create post")
 		}
 
-		_, err = s.postRepository.InsertImage(ctx, post.PostID, imageData.Height, imageData.Width, destinationKey, int(int32(displayOrder)))
+		_, err = s.postRepository.InsertImage(ctx, post.PostID, imageData.Height, imageData.Width, destinationKey, displayOrder)
 		if err != nil {
 			return errors.New("unable to create post")
 		}
@@ -114,7 +114,7 @@ func (s *PostService) GetPostById(ctx context.Context, currentUser models.Public
 		return nil, err
 	}
 
-	user, err := s.userRepository.GetUserById(ctx, int(post.UserID))
+	user, err := s.userRepository.GetUserById(ctx, post.UserID)
 	if err != nil {
 		return nil, err
 	}
@@ -198,7 +198,7 @@ func (s *PostService) GetMutualFeed(ctx context.Context, currentUser models.Publ
 	// Extract just the post IDs from the rows
 	postIDs := make([]int, len(postRows))
 	for i, row := range postRows {
-		postIDs[i] = int(row.PostID)
+		postIDs[i] = row.PostID
 	}
 
 	return s.getPostsByPostIDs(ctx, currentUser, postIDs)
@@ -238,13 +238,13 @@ func (s *PostService) AddLikeToPost(ctx context.Context, currentUser models.Publ
 		return err
 	}
 
-	if currentUser.UserID != int(post.UserID) {
+	if currentUser.UserID != post.UserID {
 		text := fmt.Sprintf("@%s liked your post.", currentUser.Username)
 		facets, err := repositories.GenerateFacets(ctx, s.userRepository, text)
 		if err != nil {
 			return err
 		}
-		err = s.notificationRepository.InsertNotification(ctx, int(post.UserID), &postId, nil, &facets, text, models.NotificationTypeLike)
+		err = s.notificationRepository.InsertNotification(ctx, post.UserID, &postId, nil, &facets, text, models.NotificationTypeLike)
 		if err != nil {
 			return err
 		}
@@ -266,7 +266,7 @@ func (s *PostService) RemoveLikeFromPost(ctx context.Context, currentUser models
 		return err
 	}
 
-	notification, err := s.notificationRepository.FindUnreadLikeNotification(ctx, int(post.UserID), postId, nil)
+	notification, err := s.notificationRepository.FindUnreadLikeNotification(ctx, post.UserID, postId, nil)
 	if err == nil && notification != nil {
 		if time.Since(notification.CreatedAt) <= 5*time.Minute {
 			err = s.notificationRepository.DeleteNotificationById(ctx, notification.NotificationID)
@@ -285,7 +285,7 @@ func (s *PostService) DeletePost(ctx context.Context, currentUser models.PublicU
 		return err
 	}
 
-	if int(post.UserID) != currentUser.UserID {
+	if post.UserID != currentUser.UserID {
 		return errors.New("unable to delete post")
 	}
 
@@ -299,23 +299,23 @@ func (s *PostService) getRelevantLikes(ctx context.Context, currentUser models.P
 	}
 
 	sort.SliceStable(likes, func(i, j int) bool {
-		return seededRandom(postId+int(likes[i].UserID)) < seededRandom(postId+int(likes[j].UserID))
+		return seededRandom(postId+likes[i].UserID) < seededRandom(postId+likes[j].UserID)
 	})
 
 	count := min(len(likes), 2)
 
 	mappedLikes := make([]models.RelevantLike, count)
-	userIDs := make([]int32, count+1)
+	userIDs := make([]int, count+1)
 	for i, like := range likes[:count] {
 		mappedLikes[i] = models.RelevantLike{
 			Username: like.Username,
-			UserID:   int(like.UserID),
+			UserID:   like.UserID,
 		}
 		userIDs[i] = like.UserID
 	}
 
 	// don't include the current user
-	userIDs[count] = int32(currentUser.UserID)
+	userIDs[count] = currentUser.UserID
 
 	hasOtherLikes, err := s.likeRepository.HasLikesFromOthers(ctx, postId, userIDs)
 	if err != nil {
@@ -373,7 +373,7 @@ func (s *PostService) GetPollDetails(ctx context.Context, currentUser models.Pub
 	voteCountMap := make(map[int]int64)
 	totalVotes := int64(0)
 	for _, voteRow := range voteTotals {
-		voteCountMap[int(voteRow.OptionIndex)] = voteRow.Count
+		voteCountMap[voteRow.OptionIndex] = voteRow.Count
 		totalVotes += voteRow.Count
 	}
 
@@ -410,14 +410,14 @@ func (s *PostService) VoteOnPoll(ctx context.Context, currentUser models.PublicU
 	}
 
 	// send notification to poll owner (unless voting on own poll)
-	if currentUser.UserID != int(post.UserID) {
+	if currentUser.UserID != post.UserID {
 		optionTitle := post.Attributes.Poll.Options[optionIndex]
 		text := fmt.Sprintf("@%s voted \"%s\" in your poll.", currentUser.Username, optionTitle)
 		facets, err := repositories.GenerateFacets(ctx, s.userRepository, text)
 		if err != nil {
 			return err
 		}
-		err = s.notificationRepository.InsertNotification(ctx, int(post.UserID), &postId, nil, &facets, text, models.NotificationTypePoll)
+		err = s.notificationRepository.InsertNotification(ctx, post.UserID, &postId, nil, &facets, text, models.NotificationTypePoll)
 		if err != nil {
 			return err
 		}
@@ -451,7 +451,7 @@ func (s *PostService) GetPostsWithTimeOffset(ctx context.Context, currentUser mo
 		}
 		postIDs = make([]int, len(postRows))
 		for i, row := range postRows {
-			postIDs[i] = int(row.PostID)
+			postIDs[i] = row.PostID
 		}
 	case FeedTypeProfile:
 		if userId == nil {
